Add unaligned to sorted when no good match is found
We used to add unaligned parts to extra, dropping important
signal that the strings don't align. We'll now add them to sorted,
then continue trying to match within slop. When we reach the end of the
strings, we treat the first remaining max_slop items as valid slop,
and add the remaining unmatched parts to sorted to count against similarity