names: Map str -> Set[Symbol] in the tagger instead of a list
This makes more sense since we only want to know what symbols a name part can mean, not how many different mappings exist for that association. Since we add every single rendition under the sun of a symbol (display, normalized, multiple languages), we get quite a lot of double mappings.

The way I noticed this is that in `Tagger`, we return an entry for each Symbol, so we end up with multiple Spans being generated. For "Ltd", that's currently two (completely equal) LLC spans.